+++ c3po/db/common/base.py
@@ -1,4 +1,5 @@
 from urllib.parse import quote_plus
+from contextlib import contextmanager
 from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
@@ -15,3 +16,15 @@ def session_factory():
 
     Base.metadata.create_all(engine)
     return _SessionFactory()
+
+@contextmanager
+def session_scope():
+    session = session_factory()
+    try:
+        yield session
+        session.commit()
+    except Exception as e:
+        raise
+        session.rollback()
+    finally:
+        session.close()
