+++ c3po/db/dao/__init__.py
@@ -1,0 +1,4 @@
+from c3po.db.dao.artist import Artist, ArtistGenre, ArtistSong
+from c3po.db.dao.genre import Genre
+from c3po.db.dao.link import Link
+from c3po.db.dao.song import Song, SongGenre
